// apps/web/src/lib/api/generated/model/badgeRequestTitle.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Petclinic MSS301 — Aggregated API
 * Aggregated from: auth, customers, vets, visits
 * OpenAPI spec version: 1.0.0
 */

export type BadgeRequestTitle =
  (typeof BadgeRequestTitle)[keyof typeof BadgeRequestTitle];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const BadgeRequestTitle = {
  ROOKIE: 'ROOKIE',
  EXPERIENCED: 'EXPERIENCED',
  MASTER: 'MASTER',
  SURGERY_EXPERT: 'SURGERY_EXPERT',
  RESEARCH_AWARD: 'RESEARCH_AWARD',
  TOP_RATED: 'TOP_RATED',
} as const;