// apps/web/src/lib/api/generated/model/badgeResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Petclinic MSS301 — Aggregated API
 * Aggregated from: auth, customers, vets, visits, billing, products
 * OpenAPI spec version: 1.0.0
 */
import type { BadgeResponseTitle } from './badgeResponseTitle';

export interface BadgeResponse {
  id?: number;
  vetId?: number;
  title?: BadgeResponseTitle;
  awardedDate?: string;
  description?: string;
}